Refactor: Update import paths for types in FormCreateStudio components
